Big refactoring, handshake messages
Refactored the plugin to use a pure socket connection
instead of twisted, this avoids confusion since the plugin
fundamentally is blocking on iterating over stdin input, it
also avoids needing a big dependency like twisted when running
the plugin in whatever environment c-lightning is running in.
Secondly, added state management of peers and the requirement
for non-directory peers to handshake with directory peers on
startup, to ensure version match and pass basic info.