Clean jmcl.py of jmbase dependency.
After this commit, the plugin jmcl.py is
no longer depending on the JM codebase for
anything so can be run outside of a JM environment.
Also the custom run() method that was created to
allow use with the twisted reactor is removed, and
replaced with the default Plugin.run().